Renamed canvas.ts to graph-editor-canvas.ts
