+++ app/components/graph-editor/math.ts
@@ -8,7 +8,7 @@
 // Imports /////////////////////////////////////////////////////////////////////
 
 
-import { point } from "./canvas";
+import { point } from "./graph-editor-canvas";
 
 
 // Constants ///////////////////////////////////////////////////////////////////
